+++ src/Description.php
@@ -14,8 +14,13 @@ use Attribute;
 #[Attribute(Attribute::TARGET_CLASS)]
 class Description
 {
+    /**
+     * @param ?array<string> $examples
+     */
     public function __construct(
-        public string $description
+        protected(set) string $description,
+        protected(set) ?string $usage = null,
+        protected(set) ?array $examples = null,
     ) {
     }
 }